+++ monks_kmodule/syscalls/write.c
@@ -39,10 +39,10 @@ asmlinkage void hooked_sys_post_write(unsigned int fd, char __user *buf, size_t 
 
 	if(IS_ERR((void *)r)){
 		i->result = "Error";
-		i->details = kasprintf(GFP_KERNEL, "Errno %zd", r);
+		i->details = kasprintf(GFP_KERNEL, "Errno %d", (int)r);
 	}else{
 		i->result = "Ok";
-		i->details = kasprintf(GFP_KERNEL, "Write %zd bytes (was requested to write %zd)", r, count);
+		i->details = kasprintf(GFP_KERNEL, "Write %d bytes (was requested to write %zd)", (int)r, count);
 	}
 
 	nl_send(i);
@@ -87,10 +87,10 @@ asmlinkage void hooked_sys32_post_write(unsigned int fd, char __user *buf, size_
 
 	if(IS_ERR((void *)r)){
 		i->result = "Error";
-		i->details = kasprintf(GFP_KERNEL, "Errno %zd", r);
+		i->details = kasprintf(GFP_KERNEL, "Errno %d", (int)r);
 	}else{
 		i->result = "Ok";
-		i->details = kasprintf(GFP_KERNEL, "Write %zd bytes (was requested to write %zd)", r, count);
+		i->details = kasprintf(GFP_KERNEL, "Write %d bytes (was requested to write %zd)", (int)r, count);
 	}
 
 	nl_send(i);
